Split `go vet` and `golint` checks into separate workflow jobs
Splitting a workflow into distinct jobs (and steps within those jobs in the case of operations that need to occur in the
same environment) makes interpreting a CI failure much easier. The check results are displayed on a per-job basis, each
job having its own logs, and the results of that job shown on a per-step basis, each step having its own folded section
in the logs. This means you can quickly identify which part of the workflow failed and get the logs specific to that.