Avoid pollution of `go.mod` file by `go:lint` task
Running the `go get` or `go list` commands from the module folder results in a change to the `go.mod` file. That change
is reverted when you run `go mod tidy`. This results in cryptic diffs that can be confusing to contributors. The problem
is avoided by running the commands from a separate folder.