Move Go linting workflow jobs to "Check Go" workflow
GitHub Actions provides a paths filter feature that allows you to restrict workflow run triggers to changes to specific
paths within the repository. This can greatly improve the efficiency of the CI system by avoiding running irrelevant
checks. The effective use of this feature requires that the purpose of a workflow be focused on a specific type of file
or project component. Even though the "Lint Code" workflow was in effect focused only on Go code, it's stated purpose was
for the linting of all code. The "Check Go" workflow is collects the linting and formatting checks specific to the
project's Go files.