Add check for out of sync go.mod to "Check Go" workflow
Changes to the go.mod manifest file or go.sum lock files can be unexpected and cryptic. The contributor may be confused
about whether to commit these to the repository, or may commit unwanted changes, causing confusion to the next
contributor when they get a mysterious diff.

To avoid this, a check is added to the "Check Go" CI workflow which fails the job if a run of `go mod tidy` results in
any diff.