Remove unnecessary variable use from `go:lint` task
The`GOLINTFLAGS` variable is only used in a single place, so it doesn't provide any benefit from the standpoint of code
duplication. I don't feel that the taskfile usability is improved at all by defining these flags in a location far
removed from the command they are used with, and it makes the taskfile more cryptic to contributors who are not
experienced with taskfiles. So I think it's better to simply remove this variable.