Add checks to tasks for shell tools` being installed
The preferred approach is for the tasks to install their specific dependencies as needed, in a manner that does not
pollute the task user's environment. However, it doesn't appear to be feasible to do this with the shfmt and ShellCheck
tools used by the `shell:format` task, so the task user is responsible for installing this tool.

In the event the task user does not have such a tool installed (or installation not in PATH), the task should provide a
helpful error message, which is provided by this change.