Move shell script-related workflow jobs to "Check Shell" workflow
GitHub Actions provides a paths filter feature that allows you to restrict workflow run triggers to changes to specific
paths within the repository. This can greatly improve the efficiency of the CI system by avoiding running irrelevant
checks. The effective use of this feature requires that the purpose of a workflow be focused on a specific type of file
or project component. Since collecting related jobs into a single workflow can be the best approach both from a
maintainer and contributor viewpoint, it makes sense to do this grouping based on component type rather than on category
of operation, as was previously done by the "Check formatting" workflow.

The "Check Shell" workflow is collects the linting and formatting checks specific to the project's Shell script files.