Add check for non-executable scripts to "Check Shell" workflow
Since Windows doesn't have the concept of an executable file mode, and shell scripts not considered executable by Linux,
etc. will work fine if a shell is installed (e.g., Git Bash), shell scripts written or worked on by Windows developers
may not be executable. For this reason, it is especially helpful for the CI to do a check on the file mode of the
project's shell scripts, failing if they are not executable.