Remove superfluous shell formatting check task
The `shell:check-formatting` task has no value to the developer, since they will be better off to just run the formatting
task. The CI also doesn't make use of this task, instead taking the superior approach of running the formatting task and
then checking for a diff. Since there is no use for this task, it should be removed.