Use formatting task to check docs and config formatting in CI
The `config:check-formatting` and `docs:check-formatting` tasks have no value to the developer, since they will be better
off to just run the formatting task. This means that the only use for this task is the CI. I think this can be better
achieved by configuring the CI to format the code and then check for a diff. This ensures that there is no possibility
for a mismatch between the formatting check and the formatting task.