Add general comment to "Check Prettier Formatting" workflow providing trigger event reference
The workings of the `workflow_dispatch` and `repository_dispatch` events is not obvious from the name. For this reason,
a comment was added to explain their working. But this information is explained well enough in GitHub's documentation, so
the workflow will be less cluttered by simply providing a reference link interested parties can easily follow to get all
the information on the events used by the workflow.