Expand usage of variables in `workflows:validate` task
This facilitates the addition of a comment indicating the repository location of the schema file in case an error is
discovered.

As for the `WORKFLOWS_DATA_PATH` variable, I'm not sure why I did that (this is being imported from the template workflow
which was developed as an independent project 6 months ago).