Enable JSON schema validator support for "format" keyword
Support for the JSON schema "format" keyword was moved to a separate package in ajv v7.0.0 (ajv-cli v4.0.0). If this package is not
installed and specified as a module via the ajv-cli command, validation against any schema that uses "format" fails, even
when the instance document is completely valid. Even though none of the JSON schemas currently in use by the templates
and CI system have a "format" keyword, there's no reason one couldn't be added at any moment, so it's safest to just add
support now.

The change to ajv was done for security purposes when used with untrusted data, which is not a concern here.