Suppress irrelevant output when running workflow validation task
By default, Task displays the commands it is running. With a simple task, this is reasonable for the sake of transparency
of what is being done. However, with more complex tasks, it clutters up the output with information that is largely
irrelevant to the task user. This is avoided by the addition of the `--silent` flag. Note that this does not suppress the
output from the commands the task runs, but only the output from Task itself.