Use standardized job name for "Check Workflows" workflow
This is the naming conventions established in the standardized template workflow.

Even though it was never realized, the original idea was for the scope of this workflow to be for linting of the
repository's configuration file, paralleling the established "Check Config File Formatting" workflow. However, that
approach leads to inefficient workflow trigger path filters. The better approach is to scope workflows to a file type. So
it is more appropriate to scope the workflow to any checks specific to GitHub Actions workflows, and the workflow name
should reflect that purpose.