Use day name in "Check Workflows" workflow's schedule trigger
At the time it was written, there was some lack of support for the use of day name abreviations in the crontab definition
for the `schedule` trigger of GitHub Actions workflows (perhaps it was the JSON schema?). Since that time, the issue was
resolved. The crontab is easier to understand with the day name in place of day number.