Use standardized job name for "Check License" workflow
This is the naming convention established in the standardized template workflow.

Even though it was never realized, the original idea was for the scope of this workflow to be for linting of the
repository's documentation files. However, that approach leads to inefficient workflow trigger path filters. The better
approach is to scope workflows to a file type. So it is more appropriate to scope the workflow to any checks specific to
the license, and the workflow name should reflect that purpose.