Don't use Task for license check
Task provides a convenient way to run common development processes locally with a single command. But I don't think the
need to check the license file is common, nor something anyone would bother running. Instead, it's something that can be
added to the repository and forgotten until the day comes that someone mucks around with the license file. That person
likely would not have been aware of the existence of the task.