Use specific linting task "Check Python" workflow
"Umbrella" tasks that run multiple related specific tasks can be convenient for the developer (though not so much in this
particular case) by allowing them to avoid typing extra commands.. However, these tasks should not be used in the CI
workflows. The reason is that more granular checks make the workflow result easier to understand in the event of a
failure.