Quote all paths in release workflows
Even though they are not necessary, it is best practices to always quote paths. Since the paths in the workflow use the
project name, it might be that quoting is necessary for a specific project and this is intended to be a template that can
apply to any project with the minimum of modification.