Add manual trigger events to "Publish Nightly Build" workflow
A situation might arise where it is desirable to immediately trigger the nightly workflow. For example:

- Testing after changes to the workflow
- Testing after changes to or affecting the secrets used by the workflow
- Testing after changes to the download server infrastructure
- Immediately pushing a project change via the nightly distribution channel

I have found that the manual workflow trigger events stay harmlessly out of my way until the time when they become very
convenient in situations like the ones I listed above.

Although I haven't found a need for the `repository_dispatch` event yet, I think its could be valuable in the event that
it was necessary to trigger the nightly on multiple projects, in which case being able to automate the trigger would be
convenient.