Report all failures of "Publish Nightly Build" workflow
Previously, the configuration of the workflow was such that a Datadog report was only submitted when the
`publish-nightly` job failed. In the event one of the other jobs failed, the `publish-nightly` job never ran, and neither
did the failure handling step inside the job.

We already had a real life event where the Arduino CLI nightly was failing due to a notarization problem, which was not
reported due to the workflow configuration.