Sanitize git log output used to provide versioning information to build commands
The Go release system uses a `git log` command to determine the commit in order to populate the
output of the application's `version` command. The expected output of the command is solely the short hash. With the
previous command, that expectation was not realized under the following conditions:

- The `log.showSignature` Git configuration option is set to true.
- The commit being built was signed (as is always the case when a commit is made via the GitHub web interface).

In this case, the signing information is shown in addition to the hash, breaking the build system.