Update website generation dependencies
NOTE: the outdated versions of mike and mkdocs-material, previously in use required an approach to creating the version
selector menu the author describes as "hacky". That "hacky" approach no longer works with the modern version of
mkdocs-material. Fortunately, this is not a problem since the material theme now has official support for versioning via
mike:
https://squidfunk.github.io/mkdocs-material/setup/setting-up-versioning/
This avoids the need for us to maintain JavaScript and CSS files.

This does result in a change in the website GUI. The version selector menu has moved from where it was previously in the
navigation panel to the title bar.