Run Go module configuration sync check also on docsgen module
The `check-config` job of the "Check Go" workflow runs a check to see whether there were any missed updates to the
`go.mod` and `go.sum` files. Previously, this was only done for the main module. But there is an additional module for
generating the website's command reference content the update of which might easily be overlooked. The CI check was not
applied to this module, which could allow its metadata to get out of date.

The use of a job matrix allows the workflow to easily accomodate the future addition of more modules to the repository by
simply adding their path to the matrix array.