Always use `docs:generate` task for documentation generation
Even though this project's generated documentation is currently limited to the CLI reference, this might not always be
so, and anyway this is intended to follow a universally applicable template.

`docs:generate` is an "umbrella" task that collects all the specific documentation generation subtasks a project has, and
so it should be used whenever the documentation needs to be generated.