Remove Python dependencies caching step
During the development phase of the previous attempt at setting up an official collection of template workflows, I
investigated dependencies caching and added it to all workflows where applicable. Afterwards, doubts were raised about
whether it provided any benefits. I don't know enough about this subject to make a call on that and I have been unable to
get any more information on the subject.

Since the caching significantly increases the complexity of the workflows, which may make them more difficult to maintain
and contribute to, I think it's best to just remove all the caching for now. I hope to eventually be able to revisit this
topic and restore caching in any workflows where it is definitely beneficial.

This particular caching step was in the source workflow in the Arduino CLI repository from the very beginning, and was my
inspiration for using caching in the first place. However, none of the other workflows with Python dependencies have
caching at this point, so having this one as an exception to the otherwise fairly standardized workflow format is
confusing.