Exclude go.mod/go.sum files from repo's spell check task recursively
The repository has a spell check as part of its infrastructure to catch typos before they can be introduced to the
content. This check should not be done on generated files like `go.mod` and `go.sum`. The codespell spell check tool was
already configured to skip those files in the root of the repository, but there is another module in `./docsgen`, and its
module configuration files must also be excluded from the check.