Move integration test runs to dedicated workflow
Due to the tests being located in a dedicated folder and written in Python, there are unique paths filter requirements
for the GitHub Actions workflow that runs the project's integration tests. For this reason, and also to enable the unit
test workflow to be applied to projects that use other integration testing approaches (i.e., none at all), it is better
to use separate workflows to run unit tests vs. integration tests.