Move generated code sync check to dedicated workflow
I have found that the approach of focusing each GitHub Actions workflows on a specific type of task can have some
significant benefits

- Improved efficiency due to being able to use a refined paths filter to prevent pointless runs of the workflow
- Easier for contributors to interpret the results in the event of a failure
- Easier to maintain
- More likely to be applicable as a "template" to arbitrary projects

This check really had no business being in the test runner workflow. It could be argued that it is suitable for inclusion
in the "Check Go" workflow, but it does have unique paths filter requirements and is also not in that workflow's
"template". So, at least for now, I am putting it in a dedicated workflow.