Fix error in `go test` command template
Task's `default` template command uses the first value if the second is not defined. In this case, the second value was a
variable defined by the workflow, so it was always defined! This made the default command and the first value useless. My
interpretation is that the intent was, as is done successfully elsewhere, to allow the task user to make customizations
by defining an environment variable from the command line which will override the default value. In order to accomplish that in a clean manner, I made the following changes:

- Remove the frivolous taskfile variable, instead adding the flags directly to the template
- Hard code the `-v` flag into the command
- Add an override variable with a more relevant name