Replace broken website publishing script link
The Python script used by the website publishing system was moved during the sync to the "template" version but the link
to it from CONTRIBUTING.md was not updated accordingly.

The reworking of the system that broke the link also removed the documentation content from that file. That content has
been moved to the documentation for the template system. Since this content is for maintainers, and is not of interest
to the average contributor, I think it is find to maintain that content in a single place in the assets repository, and
link to it there.