Use workflow variables for tool versions in all workflows
Placement of this information at the top of the file, where it is easy to find and edit, facilitates updates to the
workflows as the tool version used for project development are bumped periodically.

This is an established convention in the "template" workflows, but I missed some applications of it. I think it's most
effective when done consistently throughout the workflows.