Bump `codecov/codecov-action` workflow dependency to v2
The `codecov/codecov-action` action is used in the unit test runner workflow to upload code coverage data to codecov. The
workflow is configured to use the latest release within the specified major version series. It is necessary to manually
update the workflow any time a new major version series of an action begins, in which case a review is necessary to
determine whether the workflow is affected by the breaking changes that required the major bump.

The action is now in its 2.x version series. The cause of the major bump was the removal of some of the action's inputs.
I reviewed the changes and found that none of the inputs used in the workflow are affected so it is possible to update
without any other changes to the workflow.