Remove test suite for "siteversion" script
This script is now developed and maintained in a dedicated repository for such reusable project assets. This means that
there is no need for its test code to be checked in to this repository. I did this by accident after having failed to
follow the installation instructions carefully and am now correcting my mistake.