Make workflow path filters handle either YAML file extension
There are two file extensions in common use for YAML files: `.yaml` and `.yml`. Although this project uses `.yml`
exclusively for YAML files, standardized "template" workflows which might be applied to projects that have established
the use of the other extension. It will be most flexible if it supports both.