Consolidate convenience "umbrella" tasks
There is an established practice of providing the developer with convenience tasks which collect all the domain-specific
tasks of a given class. This has a couple of benefits:

- Provides an alternative to manually running multiple tasks
- Provides a standardized interface to the project, independent from its unique underlying details

However, I have come to the conclusion that these benefits lessen as the number of such convenience tasks increases. At
the same time, the maintenance burden increases. In the end, I think there are only three distinct classes of tasks that
can be consolidated in this manner:

- "build" (build the application)
- "check" (check for problems with the project)
- "fix" (automatically fix problems with the project)