Alphabetize tasks
I originally set out to establish some logical order to the tasks, based partly on the Arduino CLI taskfile, as that
project was used as a model, then on some indecipherable criteria of my own devizing in cases where the model did not
provide guidance. In the end, it resulted in a jumble.

Alphabetical order is completely objective and it results in a fairly logical order in the end due to the the use of
prefixes on the task names according to their domain. The exception is that the convenience "umbrella" tasks have been
left at the top (in their own alphabetical order) so that they will have the maximum visibility as the most useful tasks.
Even that is not completely inconsistent, since these don't have a namespace prefix, and so might be imagined to have a
null global namespace prefix that would be sorted first.