Replace problematic release of `github.com/oleiade/reflections` dependency
This indirect dependency had a force push that changed the checksum of its v1.0.0. This causes Dependabot to fail,
meaning it won't provide any notifications of dependency updates (nor is Dependabot so kind as to provide a notification
of failures).

There has been a new release of `github.com/oleiade/reflections` and this is slowly working its way up Arduino Lint's
supply chain, but it's still not here. So I am forcing a replacement of the module's v1.0.0 tag with the v1.0.1, which is
identical to v1.0.0. This change should be reverted once the supply chain provides v1.0.1 directly.