Fix handling of missing lib examples folder
The `rulefunction.MissingExamples` function was previously relying on the github.com/arduino/go-paths-helper IsDirCheck() method for two things

- Does examples path exist?
- Is examples path a folder?

The code made the unreasonable assumption that IsDirCheck() would return `false, nil` if the path did not exist, and an
error only if some unexpected failure occurred. That is not how IsDirCheck() is intended to work, as is clearly described
in its documentation. A bug fix to align the method's behavior with that stated in its docs caused it to correctly return
an error when the path did not exist, which caused an inappropriate panic from the `rulefunction.MissingExamples`
function when checking an examples folder that did not exist.

There is a dedicated method for checking path existence. By calling this method before `IsDirCheck()`, the code can
achieve its intended behavior of handling missing examples folders gracefully, only panicing in the event of some
unexpected failure that can't be handled without risking spurious linting results.