Add reference URL field to rule configuration
Many of the rule messages contain a URL that can be visited to find the detailed information about the problem and how
to fix it.

Moving this information to a dedicated field in the struct allows for improved formatting of the rule output as well as
enhanced capabilities for the use of this data in other applications.