Add and expand rule descriptions
The `Description` field of the rule configuration was previously only used to add supplemental information to the verbose
tool output. In this usage, it did not need to stand on its own as a complete documentation of the rule, and also was not
a high priority since the verbose output is likely not as frequently used.

A new requirement for the rule configuration data has emerged from the project to publish a list of rules. The
`MessageTemplate` data is not suitable for use in this application due to being written for use in indicating a problem
that was found, rather than a documentation of the rule. The new requirement is compatible with the existing use of the
`Description` field, so it is not necessary to add a new field. However, the existing content of the `Description field
is mostly not inadequate, so it must be expanded.

This content is now written in Markdown. Its existing use is limited to the JSON format output. Since the Markdown syntax
is very lightweight and intuitive, and any reader of the JSON format output will already have been presented with plenty
of syntax, I don't think the change has a harmful effect on that output.