Add multi-module support to project infrastructure
This projects contains two Go modules, one in the root and one in the `docsgen` subfolder. More will be added. In order
to support checks on the supplemental modules in the project subfolders, it's necessary to configure the commands to run
from their path. This is passed to the task via the GO_MODULE_PATH environment variable. If this variable is not defined,
the default root module path is used as default, preserving the previous task behavior.

The workflows use a job matrix to allow easy configuration for any number of module paths and a dedicated parallel job
for each module.