Escape Markdown markup in rule reference Brief text
The primary purpose of text of the rule configuration's `Description` field is for display in the rule reference section
of the documentation website. For this reason, it is written in Markdown, and thus can be output as is.

The situation is different with the rule configuration's `Brief` field, since it is displayed prominently in the tool
output. For this reason, the use of Markdown would not be appropriate. This text may contain incidental markup characters
that would result in unwanted formatting and thus the `Brief` field text must be escaped for display on the website.