Use different syntax style for arbitrary array elements in rule descriptions
Some of the rules must refer to data paths in configuration files such as the package index. In the case of arrays, it is
necessary to indicate that the reference applies to an arbitrary element. Prefiously, the `[]` syntax was used for this
purpose (e.g., `packages[].tools[]`). It is now changed to using `[*]` (e.g., `packages[*].tools[*]`).