Refer to the platform "extra_flags" properties in a less confusing manner
There is a convention of configuring Arduino boards platforms to provide a user interface for configuring the build
process through the addition of references and fallback empty definitions of a series of properties that follow the
format of `compiler.<pattern type>.extra_flags`, where "<pattern type>" corresponds to the compilation pattern that
references them:

- `c`
- `cpp`
- `S`
- `ar`
- `c.elf`

Previously, `x` was used as a placeholder for the component of the property name that varies in order to allow a single
rule to refer to the entire set. "<pattern type>" is a little bit more clear than "x".