Add capability to test for required subproperties in platform.txt
An optional property may have a required subproperty. Previously, the test only allowed testing for required root
properties, by removing the property entirely. In order to test for required subproperties, it is necessary to add a root
property that is missing the subproperty in addition to removing the required property.