Add utility function for parsing properties lists
Used as a convention for years, the integer terminal subproperty's use as an array syntax has now been formalized as a
"list" type. Although the properties package provides a function to parse this data type, it still takes some code
acrobatics to convert that into the map[string]interface{} type consumed by the JSON schema parser, which is a candidate
for a function.

Another "set" data type has also been added, but I don't have an immediate need for it, so will wait to add
support.