Add parent check capability to required boards.txt property helper function
Due to the complexity introduced to the data format via the custom board options system, I was not able to find any way
to define some of the boards.txt rules via the JSON schema. These were implemented in pure Go code. The requirement has
emerged to check whether a required property exists either as a terminal property or as a parent property (e.g.,
`uno.upload.tool` or `uno.upload.tool.serial`). A helper function is already present with the logic for handling the
check of whether the property exists either directly based on the board ID or as a custom board option property, but it
didn't have the parent property support, which is added here.