Move the "booleanString" schema to the general definitions file
This is an enum that contains the supported string representations of Boolean values. It was previously used only by the
boards.txt schema, and so was hosted in its file, but now there is a need for it in the platform.txt specification as
well, so it goes in the file that holds the shared definitions.