Fix properties mapping utility function handling of overlapping keys
The Arduino project configuration fields have an odd usage of the properties.Map format. Dots may sometimes indicate
nested keys, but in other cases they are merely a character in the key string. Previously, the utility function used to
recursively map to a configurable depth was not correctly handling this situation due to using the same code both for the
recursive mapping to subkeys as well as for the flat mapping at the final depth. The fix for the bug also makes the
function more efficient by avoiding calling SubTree() when it's already known there is no subtree.