Increase mapping recursion depth of platform.txt tools.* properties
The Arduino project configuration files make an unusual usage of the "properties" data format in that sometimes a `.` in
a key name indicates a nested data level, while other times it is only a character. There are cases where a completely
programmatic recursion of the properties into a fully nested structure would result in the impossibility of some keys
having both a string and a map type, which is not supported by the Go map type that holds the recursed data, nor the JSON
data format it is validated against. For this reason, it's necessary to manually configure the recursion of key levels on
a case-by-case basis and the approach is taken of recursing only when necessary.

Previously, it was possible to treat the params.verbose/quiet component of the
`tools.TOOL_NAME.ACTION_NAME.params.verbose/quiet` properties as being just key names. However, the addition of the
`tools.TOOL_NAME.upload.field.FIELD_NAME` and `tools.TOOL_NAME.upload.field.FIELD_NAME.secret` properties requires one
more level of recursion in order to deal with the fact that the FIELD_NAME component is arbitrary. This resulted in
splitting the params.verbose/quiet property, which required adjustments to their schema and rules.