Add `.gitconfig` file type to `.editorconfig`
The `.gitmodules` file defines the properties of a repository's submodules. The file automatically generated by Git
submodule commands use tabs for indentation.

It uses the same file format as the Git configuration file (e.g., `.gitconfig`). Even though the `.gitconfig` file is not
likely to be found under the repository tree, it's possible the `.editorconfig` might end up being used outside the
project specific scope so I added it to the file pattern.