Improve readability of macOS repackaging commands
Some complex and lengthy commands are used to repackage the macOS build of the Go application after notarization. Adding
some escaped line breaks and indenting the line continuations to indicate the command structure makes it easier to read
and understand.

This was already done for the other complex commands of the workflows, but was neglected for the repackaging step.