Allow use of code under later versions of GPL
The license type of Arduino Lint is hereby changed from "GPL-3.0-only" to "GPL-3.0-or-later", for the reason explained here:

https://www.gnu.org/licenses/gpl-faq.html#VersionThreeOrLater

Arduino Lint's GPL 3.0 license itself provides for either type, but the "or later" exception is only available if it is
explicitly stated:

> If the Program specifies that a certain numbered version of the GNU General Public License "or any later version"
> applies to it, you have the option of following the terms and conditions either of that numbered version or of any
> later version published by the Free Software Foundation.

This was not previously done in the Arduino Lint code base, meaning that it could not be used by code licensed under some
later version of GPL.

The previous use of the "GPL-3.0-only" license type was not a conscious decision, but simply the result of copy/pasting
the license header from another project which happened to be "GPL-3.0-only". I simply copy/pasted the license from
Arduino CLI, which is also "GPL-3.0-only".

Since Arduino has complete ownership of all code subject to this change, no external permission must be obtained to make
this change.