Add an introduction to install docs
This document provides a list of installation options. The user may pick any one option that is most convenient for them.
However, a reader might get the impression that it is a list of instructions that must be followed from top to bottom.

An introduction makes the situation immediately clear to the reader.