Refactor package index project data component parsing code
There was an inefficient approach to populating the project data for package index components, where the slices were
iterated over redundantly for each of the components they contained. At first perhaps a reasonable tradeoff for the sake
of consistency and readability, the redundancy has became more significant over time as new components are added. With
the pending addition of yet another for pluggable monitor support, it seems best to refactor now.