Fix errors in package index project data test assertions
There were some typos in the test tables data for the discoveryDependencies components of the package index project data
unit tests. Due to missing handling code for this part of the tables, the errors were not apparent. Once the handling
code was added, they caused spurious test failures, which are fixed here.