Update golden master data for ruledocsgen module unit tests
The `github.com/arduino/arduino-lint/ruledocsgen` module contained in this repository uses the
`github.com/JohannesKaufmann/html-to-markdown/escape` package to escape rule configuration data to make it safe for
inclusion in Markdown files.

Escaping has been enhanced in the `github.com/JohannesKaufmann/html-to-markdown@1.3.1` release, which caused the output
under test to no longer match the golden master.

The golden master is hereby brought into sync with the valid and expected output when using the updated dependency.