Point user links to docs for "latest" version
The documentation hosted at arduino.github.io/arduino-cli is versioned, with the URL determining which version is loaded:

- arduino.github.io/arduino-cli/0.20 - documentation content at the time of the latest release of the 0.20.x series
- arduino.github.io/arduino-cli/latest - documentation content at the time of the latest release
- arduino.github.io/arduino-cli/dev - documentation content at the tip of the default branch

Previously, some of the links displayed to users by Arduino Lint were pointing at the "dev" version of that
documentation. Since it may not apply to the release versions, or be in an unstable state, the "dev" version of the
documentation should only be used by Arduino CLI developers and beta testers. For this reason, the "latest" version will
be most appropriate for the users of Arduino Lint.