Cover all modules of project in "umbrella" tasks
For the convenience of the developer, some "umbrella" tasks have been defined, which run all the specific tasks of a basic class:

- `check`: run all automated checks for issues
- `fix`: make automated corrections for issues where possible

The project contains multiple Go modules in different folders:

- `github.com/arduino/arduino-lint`: root folder
- `github.com/arduino/arduino-lint/docsgen`: docsgen subfolder
- `github.com/arduino/arduino-lint/ruledocsgen`: ruledocsgen subfolder

`go` commands are applied to the module in the current working directory. The tasks are configured so that the working directory can be specified, defaulting to the root.

Previously, the "umbrella" tasks only ran the `go:*` tasks with the default, meaning only the
`github.com/arduino/arduino-lint` module was covered by these tasks. This is contrary to the intended purpose of these
"umbrella" tasks, which is to allow the developer to run a couple of standardized commands to find or fix all problems,
trading efficiency for convenience.