Bump `go.bug.st/relaxed-semver` to v0.9.0
At the time the dependency on `go.bug.st/relaxed-semver` was added to the project, there was not a release version
available, so it was necessary to use a "pseudo-version" instead.

Since that time, the first release of `go.bug.st/relaxed-semver`, offering the opportunity for more effective management
of the dependency:

- Greater stability through the use of release versions
- Automated update offers from Dependabot

The relevant changes made to `go.bug.st/relaxed-semver` in the interim have been the expansion of the version
constraints syntax, so it will be important for Arduino Lint to use the same version as
`arduino/libraries-repository-engine` and Arduino CLI in order to correctly lint the constraints syntax used in the
library.properties `depends` field.